pyproject.toml
[tool.poetry]
name = "highzer"
version = "0.1.0"
description = "Twitch Stream Analyzer"
authors = ["Felix Breuer <fbreuer@pm.me>"]

[tool.poetry.scripts]
highzer = "highzer.highzer:cli"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.24.0"
click = "^7.1.2"
numpy = "^1.19.1"
python-ffmpeg = "^1.0.11"
schedule = "^0.6.0"
#scipy = "^1.7.0"

moviepy = "^1.0.3"
path = "^15.1.2"

# nix module is already built
cryptography = "3.4.*"
PyYAML = "^6.0"
python-slugify = "^5.0.2"
kubernetes = "^20.13.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
